Moved `Query` and related tests to submodule `people::query`.
